+++ paralytics/utils/__init__.py
@@ -1,3 +1,4 @@
 from .mathy import *
+from .storage import *
 from .toolkits import *
 from .validation import *